+++ l10n_it_ddt/models/__init__.py
@@ -1,23 +1,13 @@
 # -*- coding: utf-8 -*-
-##############################################################################
 #
-#    Copyright (C) 2014 Abstract (http://www.abstract.it)
-#    @author Davide Corio <davide.corio@abstract.it>
-#    Copyright (C) 2014 Agile Business Group (http://www.agilebg.com)
-#    Copyright (C) 2015 Apulia Software s.r.l. (http://www.apuliasoftware.it)
-#    @author Francesco Apruzzese <f.apruzzese@apuliasoftware.it>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 #
-#    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-#
-##############################################################################
-
-
 from . import stock_picking_package_preparation
 from . import partner
 from . import account
 from . import product
 from . import stock
-from . import sale
+from . import sale_order
 from . import stock_location
 from . import delivery_carrier
 from . import ir_sequence